requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
pip==24.2
pymongo==4.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
regex==2024.11.6
sendgrid==6.11.0
setuptools==75.1.0
starkbank-ecdsa==2.2.0
tqdm==4.67.1
Werkzeug==3.1.3
wheel==0.44.0
